requirements.txt
captcha-break
captcha
numpy
matplotlib
tensorflow-gpu
pydot
pandas
